Move helper to root of command
