+++ DiscordBot/Commands/Interactive2/Graveyard/ShameSubCommand/ShameCommandHandler.cs
@@ -1,5 +1,4 @@
 using DiscordBot.Commands.Interactive2.Base.Handlers;
-using DiscordBot.Commands.Interactive2.Graveyard.Helpers;
 using DiscordBot.Common.Models.Enums;
 using Microsoft.Extensions.DependencyInjection;
 using WiseOldManConnector.Models.WiseOldMan.Enums;
