+++ DiscordBot/Commands/Interactive2/Graveyard/Shames/ShamesCommandHandler.cs
@@ -1,6 +1,5 @@
 
 using DiscordBot.Commands.Interactive2.Base.Handlers;
-using DiscordBot.Commands.Interactive2.Graveyard.Helpers;
 using DiscordBot.Common.Dtos.Discord;
 using DiscordBot.Common.Models.Enums;
 using Microsoft.Extensions.DependencyInjection;
